requirements.txt
pandas
numpy
scikit-learn
matplotlib
ollama
anthropic
openai
google-genai
openpyxl
tqdm
python-dotenv